pyproject.toml
[project]
name = "qiao-wechat"
version = "0.3.0"
description = "Python-only WeChat/XHS content workflow: WeMD-like rendering + wxAiPost-like backend + publishing records"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "SQLAlchemy>=2.0.30",
  "requests>=2.32.0",
  "python-dotenv>=1.0.1",
  "markdown>=3.6",
  "beautifulsoup4>=4.12.3",
  "lxml>=5.2.2",
  "cssselect2>=0.7.0",
  "tinycss2>=1.3.0",
  "PyYAML>=6.0.1",
  "python-multipart>=0.0.9",
  "APScheduler>=3.10.4",
  "Pillow>=10.4.0"
]

[project.scripts]
qiao-wechat = "qiao_wechat.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.28.0"
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["qiao_wechat"]
